Fix uncommitted changes to SQLite db during fatal error
Fixes #5927

This issue arose due to #5927 where newly trapped EMS errors caused a
fatal error message. However, this was within the ManageSimulation()
main loop which is part of a transaction, meaning the errors were not
committed when a fatal error happened. This addresses that by calling
commit if it is currently within a transaction during fatal error
handling.