Fix for bug and modifications to documentation
This commit corrects the bug by checking for the condition that caused
the crash and gracefully exiting out with a severe/fatal error and
helpful hints as to how to fix the problem.  The commit also contains
corrections to the documentation to better describe the input fields
and how to avoid the problem that was discovered in the code.