Clarification of the documenation regarding the 'Main Pipe' in the VRF-FluidTCtrl-HP and VRF-FluidTCtrl-HR models
As suggested by Julien Marrec, clarification of the documentation (IDD
file, I/O reference guide and engineering manual) regarding the meaning
of the 'Main Pipe' in the VRF-FluidTCtrl-HP and VRF-FluidTCtrl-HR
models.

The proposed changed are based on an email exchange with Kaiyu Sun, one
of the developer.

"The "main pipe" refers to the pipe connecting the outdoor unit to the
first branch joint. That's where the most piping loss occurs."