Correction of many problems
This commit corrects many problems with the thermal comfort model
documentation sections for the Fanger, Pierce, and KSU models as
requested by the defect description.  Equations were rechecked and
corrected when discrepancies were found.  Now, everything lines up with
what is in the code.  Lots of changes for the better.