Correction of negative cooling reporting
Bug that was resulting in cooling being negative in output.  The
cooling is not negative—it was simply picking up the fact that the
cooling panel has a heat removal impact on zones.  However, all the
output variables are clearly cooling so these should not show up as
negative.  This commit fixes that and adds a unit test.