Fixes to reduce needless max iteration warnings in BB
The radiant/convective BB units show some fluctuations that
occasionally were resulting in some max iteration warnings for those
devices.  The problem is that the radiant component is potentially
causing this lack of convergence and so providing another iteration
limiting condition makes sense here.  The new condition looks at the
non-dimensional difference between the flow max and min for the
interval halving routine.  When the difference between these two values
is less than 0.00001, there is no reason to keep iterating because the
difference between the top and bottom of the interval is now so close
as to be the same value practically.