Create and expose a new function that uses return codes instead of hard exits
Calling exit(...) in library code is problematic as it doesn't allow the calling application to handle errors gracefully or provide error
recovery and/or cleanup. This change exposes a new return-code based library API function that can be linked to code to use
error codes similar to the exit status codes but without a hard application exit.

- the old functionality of hard exit is preserved under the original function name that uses the return code version under the hood
- prevent AbortEnergyPlus() from calling exit(...)
- keep old behavior to explicitly exit from EnergyPlusPgm for those depending on that behavior