Fix issue when parsing utf8 JSON file
This was a change I originally made during the input processor work since it was causing a build error on one platform. The issue is that this is not equivalent behavior and was causing issues when parsing utf-8 even though the JSON library fully supports utf-8. If this still fails during a build on another platform then I can revert back to the old solution but increment the range variable within the if statement.