Addition of warning about people schedules
Added a warning that informs the user that if air velocity, clothing
insulation, or work efficiency schedules are specified for a people
statement but no thermal comfort model is selected, then they will get
unused schedule warnings.  The reality is that these schedules are
useless without actually picking a thermal comfort model.