Correction of Problem
The input file used the wrong pump type for a loop that contains a
two-way common pipe.  An error was thrown in the eplusout.err file as a
result.  Changes to the input file corrected this problem and the error
message is no longer present in the err file.  This is consistent with
other IceStorage*.idf files in the test suite.  Those were checked and
they do not produce the error message.