Correction of fan heat issue
Testing of the pull request showed some strange situations where the
cooling coil was coming on to deal with the fan heat.  The algorithm
was modified to take fan heat into account and only run the coil when
it was really needed and to increase the OA amount to handle the load
and fan heat if possible.