Adjust IPLV calculation
The previous IPLV calculation assumed that the PLR was equal to the loading percentage prescribed by AHRI 550-590 (% of full load capacity). The proposed chnaged recalculate the PLR at for each loading as the ratio of the load (percentage prescribded by ASHRI 550-590) to the operating capacity (function of condenser temperature and leaving chiller temperature).