Unit tests, input file change, and documentation edits
This commit includes unit tests for the new subroutine that supports
the new output variable, a change to an input file in the test suite to
show the new output variable, and some documentation edits.  The
documentation edits include new language in the IO Reference regarding
the new output variable and edits and additions to the Eng Ref about
interior moveable insulation.  The previous version of Eng Ref did not
have any information regarding moveable insulation.