Correction of windward/leeward cutoff
The cutoff between windward and leeward is currently 100 degrees.  Why
100 degrees?  It relates back to the TARP algorithm that BLAST was
based on.  However, the DOE-2 and MoWiTT convection algorithms were
supposed to be based on a cutoff of 90 degrees.  90 degrees makes much
more sense logically.  So, the point here is to change this to 90
degrees and see what happens with the results.