Unit test and Engineering Reference Change
This commit includes the unit test for the modifications to the
Windward function and an engineering reference document change as a
result of the fix.  Windward is now cutoff once the angle between the
wind and the outward normal of the surface or facing direction become
greater than 90 degrees rather than 100.  This is more logical and is
in line with more recent references (as noted in the modification to
the Engineering Reference).