Correction of Zero Value Output for Window Face Temperature
The problem noted by a user is that when requesting Surface Window
Front (or Back) Face Temperature Layer X from EnergyPlus that for a
non-BSDF window type that these variables are not defined by anything
and thus show up as zero in the output.  This fixes this by assigning
the correct value at least at the outer and inner most faces and will
not output anything for in-between layers that are not calculated.