Maintain conditions across cooling coil
Rather than simply zeroing out the load on a cooling coil when the load
is less than zero, this addresses it while the coil is being simulated
so that inlet and outlet conditions will match when the coil should be
off.  So, in the event that the cooling coil load is negative (heating)
or the air temperature or humidity ratio increase (also a sign of
heating), the coil will shut down and send the proper conditions across
to the other side of the coil.  This avoids negative cooling coil loads
and maintains the system heat balance.