Fix of additional negative cooling coil loads
Backed out a previous commit that was modifying things at the cooling
coil since that was not the original source of the problem.  The code
in unit ventilator was not taking cooling by outside air into account
and forcing the cooling coil to run when it didn’t need to run.  This
fixes the situation.  The unit test had to be modified as well as the
new function was altered to resolve this.