Solution to Reversed Layers for IZ Constructions
When interzone surfaces are used, the constructions must be reversed.
For most materials, the layers simply need to be matched in reverse
order.  However, for window glass and equivalent glass layers, some of
the properties are different on the front and back side.  So, when this
check happens and the layers are different, it needs to see if the user
has correctly reversed these material types.  In V9.1, no check is made
and an error is thrown if the materials are different even if they are
correctly reversed.