Fix epJSON parsing
This adds the internal shim capability to find "idf_max_fields" and "idf_max_extensible_fields" when using epJSON. The function getObjectItem depends on these values, however, with the recent removal of these fields from the converted epJSON files, they no longer simulate. This fixes that issue when internal code uses getObjectItem. This was not an issue with the code that uses epJSON directly.