Correction to address review comment
The method used to decide when the component loads pulse happens was changed as a result of the initial review.  Now, it will only do the pulse on the first day of a particular simulation or on all days when the user has selected to have a weather file serve as the design environment.