Correction of various errors in document
A variety of problems were noted in Tips and Tricks.  Some of these were web sites that were no longer valid.  Others were glaring formatting issues.  In some cases, minor language edits were made (like where a chapter was totally blank).  While this was not a full clean-up of the document, it looks much better and all of the hyperlinks should now work and point to the correct address.