Unit Test and Documentation
A unit test for the bug fix was created based on the new subroutine that was developed as part of this work.  In addition, clarifying documentation was added to the Input Output Reference so that it is clear what type of surface is and is not allowed for the modeling of pools in EnergyPlus.