Change of variable name
One of the review comments requested that a name change take place for a variable related to the original bug fix.  Since the previous name only referenced radiant surfaces or ventilated slabs and not also pools, there could have been some confusion for future developers.  The name has been updated to reflect all three.  Basically, any particular surface can only be a radiant system, a ventilated slab, or a pool, but not multiples of those things at the same time.