Refactoring of Radiant data structs
First step in refactoring some of the radiant system code.  This involves redoing the definitions in the .hh file.  A few changes were also needed in the .cc file as a result of a minor change.  The goal is to also associate some of the subroutines in this file with some of the specific structs as well to make the code here much more OO before doing a lot of work on the controls issues (some minor coding did already take place).