Addition of Test IDFs
Addition of various IDFs that test the new surface controls for constant flow and electric low temperature radiant systems.  This includes changes to the IDD to get the controls to be allowed as options in the input file and to the CMake file for the CI run.  The results for each of the new input files are expected though it points to the need for additional documentation to explain some of the switching that takes place in a constant flow system when it is very close to the "deadband" and why it turns off at certain times.