Implementation of Running Mean Temperature
This commit includes changes needed to implement the running mean outside air temperature control in EnergyPlus.  It also includes changes to the IDD and various IDFs that use the constant flow system.  Finally, one unit test had to be modified to deal with a change in the variables sent to a subroutine.