Unit Tests, Docs, Transition, New IDF
This commit includes a variety of files to complete the implementation of the running mean outdoor air temperature control for constant flow radiant systems.  What's included in this commit: a new unit test, the expansion of an existing unit test, documentation in the IO Ref, transition changes for the new input parameter, and a new input file which shows that the control is operating as expected.