Fix for Maximum Flow Limit
This fix keeps the flow fraction for a variable flow system at a maximum of 1.0 and also accomodates the fraction of operation for an electric system to also be limited to 1.0 through the same subroutine.  This will likely cause some diffs in the output for the variable flow radiant systems which can then be investigated.