Refactor NFP mark sample IDF and IDD changes
The NFP mentions proposals for new keys for the
GroundHeatExchanger:System. This would result in new lines in IDF and
IDD sections. Sample IDF and IDD are presented with proposed changes.
It is helpful for the reader to quickly be able to distinguish the
new lines.

The code blocks in markdown do not allow for highlighting (e.g. bold,
underline, etc.). Additionally, html font highlighting did not appear
to render in the .md file on github. The changed lines were placed
below, and were intended to be highlighted in red (rendering md in atom
offline looked fine), but were not. Therefore, the proposed changed
lines now begin with a ">" to show distinction.