Correction of code issue
This commit corrects a problem in the code that occurred for the case where the refrigeration mixing was used with contaminant modeling.  When a generic contaminant was being modeled, it led to inappropriate results in both the CO2 and the generic contaminant results.  This has been corrected.  An existing input file was modified to capture this case, and a unit test to show that this is working properly also now exists.