Correct Divide by Zero in Simple Electric Storage
Corrected the code and the IDD to avoid a divide by zero in the simple (bucket) storage model of the electric load center.  Before the fix, a zero value was allowed for both the charging and discharging efficiencies.  These both led to a divide by zero as shown in a user file.  This has been corrected.