Correct CondFDCrash Fix and Unit Test
This commit includes the fix and the unit test for defect #8919.  Under a particular condition (CondFD used, interzone surface that has another zone as the outside boundary condition causing E+ to create it internally, and the user supplying the reversed construction but it isn't used by any user defined surface), a variable was not being set (IsUsed) for the reversed construction causing a crash (array index of zero).  This is now fixed and this particular case trapped by the new unit test.