Water Thermal Tank Fix Revision
One of the changes that was related to this problem but not the ultimate source was causing other issues.  So, this changes was backed out because it is not needed to fix the problem and it was uncertain if this was really necessary.  This fix should work and eliminate some of the problems running the test suite.