8154 Fix for Autocalculate View Factor with Light Shelf
This contains the fix for the issue and the unit test.  When using autosizing for the view factors of a window that also has a light shelf, the view factors to ground and sky were set to values based on tilt.  These two values ended up being equal to 1.0 when combined.  However, then the light shelf calculated a view factor and added that to the other values resulting in the sum of the view factors being greater than zero.  This is now fixed.